requirements.txt
Django~=2.0.6
djangorestframework
django-cors-middleware
django-cors-headers
